Add two more boundary conditions to stop recrawl
The Twitter API appears to not return the full history of lists, so if
it’s been a month and many messages have been posted since then it’s
possible to only get something like 1000 including retweets.

This was causing backfill to go into an infinite loop on my test
instance. To resolve this, the following checks have been added in this
CL:
- if only one tweet was crawled this cycle, no need to do backfill
- count the total number of backfill requests made and cap after a
certain amount